requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
supabase==1.0.3
selenium==4.15.2
beautifulsoup4==4.12.2
feedgenerator==2.1.0
webdriver_manager==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0 